% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABL1_homologous.R
\docType{data}
\name{ABL1_homologous}
\alias{ABL1_homologous}
\title{Information about 451 sequences homologous to ABL1.}
\format{
A data frame with 451 rows and 127 variables:
}
\source{
\url{https://www.ebi.ac.uk/Tools/hmmer/search/}
}
\usage{
data(ABL1_homologous)
}
\description{
A dataset containing 451 sequences homologous to ABL1, a
cytoplasmic and nuclear protein tyrosine kinase that has
been implicated in processes of cell differentiation, cell
division, cell adhesion and stress response such as DNA repair.
}
\details{
They were found using PHMMER, and contain information on the
taxonomy of the species to which they belong, theoretical
physical and chemical properties calculated on the basis of
the sequence and other variables returned by HMMER.
}
\section{Sequence data structure}{

Variables beginning with "hits." have been extracted from the hash
corresponding to the sequences. Remember that there are as many rows as
there are domains identified, not per sequence. Therefore, these
variables will be duplicated, for example, for a sequence in which 2
domains are identified.
\itemize{
\item \code{hits.name}: 	Name of the target (sequence for phmmer/hmmsearch,
HMM for hmmscan).
\item \code{hits.acc}: 	Accession of the target.
\item \code{hits.acc2}: 	Secondary accession of the target.
\item \code{hits.id}: 	Identifier of the target.
\item \code{hits.desc}: 	Description of the target.
\item \code{hits.score}: 	Bit score of the sequence (all domains, without
correction)
\item \code{hits.pvalue}: 	P-value of the score.
\item \code{hits.evalue}: 	E-value of the score.
\item \code{hits.nregions}: 	Number of regions evaluated.
\item \code{hits.nenvelopes}: 	Number of envelopes handed over for domain
definition, null2, alignment, and scoring.
\item \code{hits.ndom}: 	Total number of domains identified in this sequence.
\item \code{hits.nreported}: 	Number of domains satisfying
reporting thresholding.
\item \code{hits.nincluded}: 	Number of domains satisfying
inclusion thresholding.
\item \code{hits.taxid}: 	The NCBI taxonomy identifier of the
target (if applicable).
\item \code{hits.species}: 	The species name of the target
(if applicable).
\item \code{hits.kg}: 	The kingdom of life that the target belongs
to - based on placing in the NCBI taxonomy tree (if applicable).
}
}

\section{Domain data structure}{

Variables beginning with "domains." have been extracted from
the hash domains.
\itemize{
\item \code{domains.ienv}: 	Envelope start position.
\item \code{domains.jenv}: 	Envelope end position.
\item \code{domains.iali}: 	Alignment start position.
\item \code{domains.jali}: 	Alignment end position.
\item \code{domains.bias}: 	null2 score contribution.
\item \code{domains.oasc}: 	Optimal alignment accuracy score.
\item \code{domains.bitscore}: 	Overall score in bits, null corrected,
if this were the only domain in seq.
\item \code{domains.cevalue}: 	Conditional E-value based on the
domain correction.
\item \code{domains.ievalue}: 	Independent E-value based on the
domain correction.
\item \code{domains.is_reported}: 	1 if domain meets reporting
thresholds.
\item \code{domains.is_included}: 	1 if domain meets inclusion
thresholds.
\item \code{domains.alimodel}: 	Aligned query consensus sequence phmmer
and hmmsearch, target hmm for hmmscan.
\item \code{domains.alimline}: 	Match line indicating identities,
conservation +’s, gaps.
\item \code{domains.aliaseq}: 	Aligned target sequence for phmmer
and hmmsearch, query for hmmscan.
\item \code{domains.alippline}: 	Posterior probability annotation.
\item \code{domains.alihmmname}: 	Name of HMM (query sequence for phmmer,
alignment for hmmsearch and target hmm for hmmscan).
\item \code{domains.alihmmacc}: 	Accession of HMM.
\item \code{domains.alihmmdesc}: 	Description of HMM.
\item \code{domains.alihmmfrom}: 	Start position on HMM.
\item \code{domains.alihmmto}: 	End position on HMM.
\item \code{domains.aliM}: 	Length of model.
\item \code{domains.alisqname}: 	Name of target sequence (phmmer, hmmscan)
or query sequence(hmmscan).
\item \code{domains.alisqacc}: 	Accession of sequence.
\item \code{domains.alisqdesc}: 	Description of sequence.
\item \code{domains.alisqfrom}: 	Start position on sequence.
\item \code{domains.alisqto}: 	End position on sequence.
\item \code{domains.aliL}: 	Length of sequence.
}
}

\section{Taxonomic data structure}{

If annotate_taxonomic = "local" or annotate_taxonomic = "remote" is selected,
as many columns as available taxonomic categories will be
included. All of them will have in common that they will start with "taxa".
}

\section{Theoretical physicochemical properties data structure}{

If calculate_physicochemical_properties = TRUE the following
columns will be included. All of them will have in common that
they will start with "properties".
\itemize{
\item \code{properties.molecular.weight}: sum of the masses of each
atom constituting a molecule (Da) using the same formulas
and weights as ExPASy's.
\item \code{properties.charge}: net charge of a protein sequence based on the
Henderson-Hasselbalch equation using Lehninger pKa scale.
\item \code{properties.pI}:  isoelectric point calculated as in EMBOSS PEPSTATS.
\item \code{properties.mz}: mass over charge ratio (m/z) for peptides,
as measured in mass spectrometry.
\item \code{properties.aIndex}: aliphatic index of a protein. The aindex
is defined as the relative volume occupied by aliphatic
side chains (Alanine, Valine, Isoleucine, and Leucine). It may
be regarded as a positive factor for the increase of
thermostability of globular proteins.
\item \code{properties.boman}: potential protein interaction index.
The index is equal to the sum of the solubility values for
all residues in a sequence, it might give an overall estimate
of the potential of a peptide to bind to membranes or other
proteins as receptors, to normalize it is divided by the number
of residues. A protein have high binding potential
if the index value is higher than 2.48.
\item \code{properties.hydrophobicity}: GRAVY hydrophobicity index of
an amino acids sequence using KyteDoolittle hydophobicity scale.
\item \code{properties.instaIndex}: Guruprasad's instability index.
This index predicts the stability
of a protein based on its amino acid composition, a protein
whose instability index is smaller than 40 is predicted as
stable, a value above 40 predicts that the protein may be unstable.
\item \code{properties.STYNQW}: Percentage of amino acids (S + T + Y + N + Q + W)
\item \code{properties.Tiny}: Percentage of amino acids (A + C + G + S + T)
\item \code{properties.Small}: Percentage of amino acids
(A + B + C + D + G + N + P + S + T + V)
\item \code{properties.Aliphatic}: Percentage of amino acids (A + I + L + V)
\item \code{properties.Aromatic}: Percentage of amino acids (F + H + W + Y)
\item \code{properties.Non-polar}: Percentage of amino acids
(A + C + F + G + I + L + M + P + V + W + Y)
\item \code{properties.Polar}: Percentage of amino acids
(D + E + H + K + N + Q + R + S + T + Z)
\item \code{properties.Charged}: Percentage of amino acids
(B + D + E + H + K + R + Z)
\item \code{properties.Basic}: Percentage of amino acids (H + K + R)
}
}

\keyword{datasets}
