% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fullseq_to_HMMER_tbl.R
\name{add_fullseq_to_HMMER_tbl}
\alias{add_fullseq_to_HMMER_tbl}
\title{Add fullseq_fasta information information to an \code{AnnotatedDataFrame}
by reading fasta files.}
\usage{
add_fullseq_to_HMMER_tbl(HMMER_tidy_tbl, fasta_files = NULL, N.TRIES = 2)
}
\arguments{
\item{HMMER_tidy_tbl}{An \code{AnnotatedDataFrame} obtained through
\code{extract_from_HMMER_data_tbl}.}

\item{fasta_files}{A character vector containing either the download urls
(found in the url.fullfasta column) of the files or the path to the files.}

\item{N.TRIES}{An integer with the maximum number of attempts to read/download the file.}
}
\value{
An \code{AnnotatedDataFrame} with a new column with the fasta sequences.
}
\description{
Add fullseq_fasta information information to an \code{AnnotatedDataFrame}
by reading fasta files.
}
\examples{
phmmer_tbl <- search_phmmer(
  seqs = "MTEITAAMVKELRTGAGMMDCKN",
  seq_names = "1efu_B",
  dbs = "pdb",
  verbose = FALSE)
fullfasta_HMMER_tbl <- phmmer_tbl\%>\%
    extract_from_HMMER_data_tbl() \%>\%
    add_fullseq_to_HMMER_tbl(phmmer_tbl$fullfasta.url)
fullfasta_HMMER_tbl
}
