% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxa_to_hmmer_tbl.R
\name{add_sequences_to_hmmer_tbl}
\alias{add_sequences_to_hmmer_tbl}
\title{Add sequences from one of the output files from HMMER
server to a Data Frame obtained by searching in HMMER.}
\usage{
add_sequences_to_hmmer_tbl(data, extension = "fullfasta", max_times = 3)
}
\arguments{
\item{data}{A Data Frame obtained with HMMERutils.}

\item{extension}{A one-element character vector with
either "fullfasta" or "fasta".}

\item{max_times}{An integer with the maximum number of
trials before throwing an error.}
}
\value{
A DataFrame with a new column named "hits.fullfasta" or
"hits.fasta" with the sequences.
}
\description{
Add sequences from one of the output files from HMMER
server to a Data Frame obtained by searching in HMMER.
}
