% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxa_to_HMMER_tbl.R
\name{add_taxa_to_HMMER_tbl}
\alias{add_taxa_to_HMMER_tbl}
\title{Add taxonomic information to an HMMER_tidy_tbl using \code{annotate_with_NCBI_taxid}.}
\usage{
add_taxa_to_HMMER_tbl(HMMER_tidy_tbl, mode = "remote", rank_vc = NULL)
}
\arguments{
\item{HMMER_tidy_tbl}{An \code{AnnotatedDataFrame} obtained through
\code{extract_from_HMMER_data_tbl}.}

\item{mode}{Either "local" or "remote". If "local" you will use a local
database instead of remote resources.  You will not have to download the
database but it is slower.}

\item{rank_vc}{A character vector containing the desired taxonomic ranks.
If empty, all available taxonomic ranges will be retrieved.}
}
\value{
An \code{AnnotatedDataFrame} with new taxonomic parameters.
}
\description{
Add taxonomic information to an HMMER_tidy_tbl using \code{annotate_with_NCBI_taxid}.
}
\examples{
xml_path <- system.file(
    "/extdata/xml_example.xml",
    package = "HMMERutils"
)
taxa_HMMER_tbl <- read_hmmer_from_xml(xml_path, algorithm = "phmmer")\%>\%
    extract_from_HMMER_data_tbl() \%>\%
    add_taxa_to_HMMER_tbl(mode = "local")
taxa_HMMER_tbl
}
