% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_hmmer.R
\name{bind_and_unnest}
\alias{bind_and_unnest}
\title{Aux function used to bind and unnest wider the new list column.}
\usage{
bind_and_unnest(data, old.column, new.column)
}
\arguments{
\item{data}{Dataframe whose column is going to be splitted.}

\item{old.column}{Column to remove.}

\item{new.column}{Column to add.}
}
\value{
A DataFrame with columns \code{column} splitted into
several columns.
}
\description{
Aux function used to bind and unnest wider the new list column.
}
