% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_physicochemical_properties.R
\name{calculate_physicochemical_properties}
\alias{calculate_physicochemical_properties}
\title{Calculate EMBOSS-inspired theoretical physicochemical properties using the \code{Peptides} library}
\usage{
calculate_physicochemical_properties(seqs)
}
\arguments{
\item{seqs}{A character vector with protein sequences or an object that can be coerced to that.}
}
\value{
A DataFrame
}
\description{
Calculate EMBOSS-inspired theoretical physicochemical properties using the \code{Peptides} library
}
\section{Physicochemical Properties data structure}{

This documentation is based on the following documentation:
https://github.com/dosorio/Peptides/
\itemize{
\item \code{molecular.weight}: sum of the masses of each atom constituting a
molecule (Da) using the same formulas and weights as ExPASy's.
\item \code{charge}: net charge of a protein sequence based on the
Henderson-Hasselbalch equation using Lehninger pKa scale.
\item \code{pI}:  isoelectric point calculated as in EMBOSS PEPSTATS.
\item \code{mz}: mass over charge ratio (m/z) for peptides, as measured in mass spectrometry.
\item \code{aIndex}: aliphatic index of a protein. The aindex is defined as the
relative volume occupied by aliphatic side chains (Alanine, Valine,
Isoleucine, and Leucine). It may be regarded as a positive factor
for the increase of thermostability of globular proteins.
\item \code{boman}: potential protein interaction index . The index is equal to the
sum of the solubility values for all residues in a sequence, it might give
an overall estimate of the potential of a peptide to bind to membranes or
other proteins as receptors, to normalize it is divided by the number of
residues. A protein have high binding potential if the index value is
higher than 2.48.
\item \code{hydrophobicity}: GRAVY hydrophobicity index of an amino acids sequence using
KyteDoolittle hydophobicity scale.
\item \code{instaIndex}: Guruprasad's instability index. This index predicts the stability
of a protein based on its amino acid composition, a protein whose instability
index is smaller than 40 is predicted as stable, a value above 40 predicts
that the protein may be unstable.
\item \code{STYNQW}: Percentage of amino acids (S + T + Y + N + Q + W)
\item \code{Tiny}: Percentage of amino acids (A + C + G + S + T)
\item \code{Small}: Percentage of amino acids (A + B + C + D + G + N + P + S + T + V)
\item \code{Aliphatic}: Percentage of amino acids (A + I + L + V)
\item \code{Aromatic}: Percentage of amino acids (F + H + W + Y)
\item \code{Non-polar}: Percentage of amino acids (A + C + F + G + I + L + M + P + V + W + Y)
\item \code{Polar}: Percentage of amino acids (D + E + H + K + N + Q + R + S + T + Z)
\item \code{Charged}: Percentage of amino acids (B + D + E + H + K + R + Z)
\item \code{Basic}: Percentage of amino acids (H + K + R)
\item \code{Acidic}: Percentage of amino acids (B + D + E + Z)
}
}

