% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMER.R
\name{create_download_url_for_hmmer}
\alias{create_download_url_for_hmmer}
\title{Get the HMER URL for the output files}
\usage{
create_download_url_for_hmmer(uuid, format)
}
\arguments{
\item{uuid}{unique identifier for each result in HMMER.}

\item{format}{a string with the file extension ("fullfasta", "json", "xml"...)}
}
\description{
Get the HMER URL for the output files
}
