% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMER.R, R/utils.R
\name{download_file}
\alias{download_file}
\title{Download a file using a temp file (it should work in every OS).}
\usage{
download_file(url)

download_file(url)
}
\arguments{
\item{url}{a string with an URL.}
}
\description{
Download a file using a temp file (it should work in every OS).

Download a file using a temp file (it should work in every OS).
}
