% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_HMMER_data_tbl.R
\name{extract_from_HMMER_data_tbl}
\alias{extract_from_HMMER_data_tbl}
\title{A function that extracts information from an \code{AnnotatedDataFrame}
obtained through the \code{search} family of functions or \code{read_hmmer_from_xml}
and returns a Tidy \code{AnnotatedDataFrame}.}
\usage{
extract_from_HMMER_data_tbl(HMMER_data_tbl)
}
\arguments{
\item{HMMER_data_tbl}{An \code{AnnotatedDataFrame} obtained through the \code{search}
family of functions or \code{read_hmmer_from_xml}.}
}
\value{
An \code{AnnotatedDataFrame}, consisting of 2 parts, a tidy DataFrame
with one row for each domain identified in the sequences  and the metadata
associated to this DataFrame. In case you want to work with sequences you
could group by hits.name to work only with sequence hits.
}
\description{
A function that extracts information from an \code{AnnotatedDataFrame}
obtained through the \code{search} family of functions or \code{read_hmmer_from_xml}
and returns a Tidy \code{AnnotatedDataFrame}.
}
\examples{
xml_path <- system.file(
    "/extdata/xml_example",
    package = "HMMERutils"
)
data <- read_hmmer_from_xml(xml_path) \%>\%
    extract_from_HMMER_data_tbl()
}
