% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hmmer.R
\name{filter_hmmer}
\alias{filter_hmmer}
\title{Filter dataframe depending on a column and a threshold.}
\usage{
filter_hmmer(data, threshold = 5e-04, by = "hits.evalue")
}
\arguments{
\item{data}{Dataframe which is going to be filtered.}

\item{threshold}{Value used to filter the dataset.}

\item{by}{Column which we are going to use to filter the dataset.}
}
\value{
A DataFrame filtered.
}
\description{
Filter dataframe depending on a column and a threshold.
}
\examples{
data(phmmer_2abl)
filter_hmmer(
    data = phmmer_2abl,
    threshold = 0.0005,
    by = "hits.evalue"
)
}
