% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMER.R
\name{fix_evalue_column}
\alias{fix_evalue_column}
\title{Fix E-value column which sometimes can be a list, a character vector
or a numeric vector.}
\usage{
fix_evalue_column(column)
}
\arguments{
\item{column}{E.value column}
}
\description{
Fix E-value column which sometimes can be a list, a character vector
or a numeric vector.
}
