% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMER.R
\name{format_AAStringSet_into_hmmer_string}
\alias{format_AAStringSet_into_hmmer_string}
\title{Converts an AAStringSet into plain text so
HMMER can read it.}
\usage{
format_AAStringSet_into_hmmer_string(AAStringSet)
}
\arguments{
\item{AAStringSet}{an AASTRingSet from Biostrings.}
}
\description{
Converts an AAStringSet into plain text so
HMMER can read it.
}
