% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMER.R
\name{get_api_search_url}
\alias{get_api_search_url}
\title{Get the HMER URL for each algorithm.}
\usage{
get_api_search_url(algorithm)
}
\arguments{
\item{algorithm}{a string with the algorithm name.}
}
\description{
Get the HMER URL for each algorithm.
}
