% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer_evalues_cleveland_dot_plot.R
\name{hmmer_evalues_cleveland_dot_plot}
\alias{hmmer_evalues_cleveland_dot_plot}
\title{A Cleveland dot plot with the evalues of the domains and sequences.}
\usage{
hmmer_evalues_cleveland_dot_plot(HMMER_tidy_tbl, threshold = 0.01)
}
\arguments{
\item{HMMER_tidy_tbl}{An \code{AnnotatedDataFrame} obtained through
\code{extract_from_HMMER_data_tbl}.}

\item{threshold}{A numeric vector of one with a maximum allowed value
of evalue to draw a line.}
}
\value{
A ggplot2 object.
}
\description{
A Cleveland dot plot with the evalues of the domains and sequences.
}
\examples{
data("example_phmmer")
hmmer_evalues_cleveland_dot_plot(
    example_phmmer,
    threshold = 0.001
)

}
