% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_alignment_sequence_identity.R
\name{pairwise_alignment_sequence_identity}
\alias{pairwise_alignment_sequence_identity}
\title{Calculate the percentage of pairwise sequence identity}
\usage{
pairwise_alignment_sequence_identity(
  seqs,
  aln_type = "global",
  pid_type = "PID1",
  allow_parallelization = NULL
)
}
\arguments{
\item{seqs}{A named character vector to convert into a \code{Biostrings::AAStringSet} or a \code{Biostrings::AAStringSet} with the sequences of interest.}

\item{aln_type}{A character vector of one containing the alignment type.
Possible options are "global" (Needleman-Wunsch),"local" (Smith-Waterman) and "overlap".}

\item{pid_type}{A character vector of one containing the definition of percent
sequence identity. Possible options are "PID1", "PID2", "PID3" and "PID4".}

\item{allow_parallelization}{A character vector of one, by default NULL.
If you want to parallelize the alignment of the sequences, speeding up the
process, select \code{multisession} or \code{multicore}.}
}
\value{
A DataFrame of subclass \code{pairwise_sequence_identity}, so that it has associated S3 methods..
}
\description{
Calculate the percentage of pairwise sequence identity
}
\section{Visualize}{

The \code{plot} method can be called to visualize either a histogram or a default
heatmap. Refer to the "examples" section.
}

\section{Alignment types}{

\itemize{
\item \code{global}: align whole strings with end gap penalties.
\item \code{local}: align string fragments.
\item \code{overlap}: align whole strings without end gap penalties.
}
}

\section{Percent sequence identity}{

\itemize{
\item \code{PID1}: 100 * (identical positions) / (aligned positions + internal gap positions).
\item \code{PID2}: 100 * (identical positions) / (aligned positions).
\item \code{PID3}: 100 * (identical positions) / (length shorter sequence).
\item \code{PID4}: 100 * (identical positions) / (average length of the two sequences).
}
}

\examples{
data(ABL1_homologous)
pairwise.per <- pairwise_alignment_sequence_identity(
    seqs = ABL1_homologous$hits.fullseq.fasta[1:4],
    aln_type = "overlap",
    pid_type = "PID2"
)
plot(pairwise.per)
plot(pairwise.per, type = "heatmap")

}
