% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_alignment_sequence_identity.R
\name{pairwise_sequence_identity_heatmap}
\alias{pairwise_sequence_identity_heatmap}
\title{Plot a heatmap with the pairwise identity percentages.}
\usage{
pairwise_sequence_identity_heatmap(object, annotation = NULL)
}
\arguments{
\item{object}{A pairwise_sequence_identity.}

\item{annotation}{A character vector with as many elements as sequences.}
}
\value{
A ggplot object
}
\description{
Plot a heatmap with the pairwise identity percentages.
}
\examples{
data(pairwise_identities_ABL1_homologous)
pairwise_sequence_identity_heatmap(pairwise_identities_ABL1_homologous)
}
