% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fasta_header.R
\name{parse_fasta_headers}
\alias{parse_fasta_headers}
\title{Extracts the sequence identifier from the recognized fasta sequence headers.}
\usage{
parse_fasta_headers(headers)
}
\arguments{
\item{headers}{A vector of characters}
}
\value{
A vector of characters
}
\description{
Extracts the sequence identifier from the recognized fasta sequence headers.
}
\examples{
headers <- c(
    "gb|M73307|AGMA13GT",
    "gb|M73307|AGMA13GT",
    "emb|CAM43271.1|",
    "sp|P01013|OVAX_CHICK description"
)
parse_fasta_headers(headers)
}
