% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_NCBI_FASTA_sequence_identifiers.R
\name{parse_NCBI_FASTA_sequence_identifiers}
\alias{parse_NCBI_FASTA_sequence_identifiers}
\title{Extracts the sequence identifier from the recognized fasta sequence headers.}
\usage{
parse_NCBI_FASTA_sequence_identifiers(headers)
}
\arguments{
\item{headers}{A vector of characters}
}
\value{
A vector of characters
}
\description{
Extracts the sequence identifier from the recognized fasta sequence headers.
}
\examples{
headers <- c(
    "gb|M73307|AGMA13GT",
    "gb|M73307|AGMA13GT"
)
parse_NCBI_FASTA_sequence_identifiers(headers)
}
