% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMER.R
\name{parse_results_into_tbl}
\alias{parse_results_into_tbl}
\title{Parses results from HMMER into a tibble.}
\usage{
parse_results_into_tbl(results)
}
\arguments{
\item{results}{list with results (@seealso HMMERutils::post_query())}
}
\description{
Parses results from HMMER into a tibble.
}
