% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMER.R
\name{post_query}
\alias{post_query}
\title{Post query to HMMER and returns the httr response.}
\usage{
post_query(query)
}
\arguments{
\item{query}{a query list (@seealso HMMERutils::params_into_query_list())}
}
\description{
Post query to HMMER and returns the httr response.
}
