% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmmer_from_json.R
\name{read_hmmer_from_json}
\alias{read_hmmer_from_json}
\title{Read the results of a search with HMMER from an json file.}
\usage{
read_hmmer_from_json(file)
}
\arguments{
\item{file}{A a character vector the containing filepaths to
XML file downloaded from HMMER.}
}
\value{
A Data Frame with the results
}
\description{
Read the results of a search with HMMER from an json file.
}
\examples{
read_hmmer_from_json(
    file = "inst/extdata/phmmer_2abl.xml"
)
}
