% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmmer_from_xml.R
\name{read_hmmer_from_xml}
\alias{read_hmmer_from_xml}
\title{Read the results of a search with HMMER from an XML file.}
\usage{
read_hmmer_from_xml(xml_file_paths, algorithm = "default")
}
\arguments{
\item{xml_file_paths}{A a character vector the containing filepaths to
XML file downloaded from HMMER.}

\item{algorithm}{character vector of one containing the algorithm that was
used. It can be \code{phmmer}, \code{hmmscan}, \code{hmmsearch}, \code{jackhammer}.}
}
\value{
An \code{AnnotatedDataFrame}, consisting of 2 parts, a nested DataFrame
with the search hashes  and the metadata associated to this DataFrame.
Although all available results are available here, we recommend using the
\code{extract_from_HMMER_data_tbl} function to preprocess the data. If not
indicated, the data will be processed as if it were a sequence search.
}
\description{
Read the results of a search with HMMER from an XML file.
}
\examples{
xml_path <- system.file(
    "/extdata/xml_example.xml",
    package = "HMMERutils"
)
read_hmmer_tbl <- read_hmmer_from_xml(xml_path, algorithm = "phmmer")
read_hmmer_tbl
Biobase::varMetadata(read_hmmer_tbl)
}
