% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMER.R
\name{search_in_hmmer}
\alias{search_in_hmmer}
\title{Post a query into HMMER for different algorithms and
returns a tibble with the results.}
\usage{
search_in_hmmer(...)
}
\arguments{
\item{\\dots}{\itemize{
\item algorithm algorithm to use (phmmer, hmmscan, hmmsearch or jackhmmer)
\item seq protein sequence as a string or an alignment
(@seealso HMMERutils::format_AAStringSet_into_hmmer_string())
\item hmmdb a string with a hmmdb (for hmmscan)
\item seqdb a string with seqdb (for phmmer, hmmsearch or jackhmmer)
\item timeout_in_seconds an integer with the number of
seconds to wait before exits.
}}
}
\description{
Post a query into HMMER for different algorithms and
returns a tibble with the results.
}
