% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_jackhmmer.R
\name{search_jackhmmer}
\alias{search_jackhmmer}
\title{Perform an iterative search of a protein sequence or an alignment against
a protein sequence database.}
\usage{
search_jackhmmer(
  seqs = NULL,
  alns = NULL,
  seq_or_aln_names = NULL,
  dbs = "swissprot",
  fullseqfasta = TRUE,
  verbose = TRUE,
  N.TRIES = 1
)
}
\arguments{
\item{seqs}{A character vector containing the sequences of the query or
any other object that can be converted to that.}

\item{alns}{A Biostrings::AAMultipleAlignment or a list of
Biostrings::AAMultipleAlignment.}

\item{dbs}{A character vector containing the target databases. Frequently
used databases are \code{swissprot}, \code{uniprotrefprot}, \code{uniprotkb}, \code{ensembl},
\code{pdb} and \code{alphafold}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}

\item{N.TRIES}{An integer specifying the number of attempts before
an error occurs.}
}
\value{
An \code{AnnotatedDataFrame}, consisting of 2 parts, a nested DataFrame
with the search hashes, the download links of all available files and
of the HMMER page where the results are hosted, and the metadata
associated to this DataFrame. Although all available results are available
here, we recommend using the \code{extract_from_HMMER_data_tbl} function
to preprocess the data.
}
\description{
Perform an iterative search of a protein sequence or an alignment against
a protein sequence database.
}
\examples{
jackhmmer_tbl_using_seq <- search_jackhmmer(
   seqs = "MTEITAAMVKELRTGAGMMDCKN",
  seq_or_aln_names = "1efu_B",
  dbs = "pdb",
  verbose = FALSE)
jackhmmer_tbl_using_seq
Biobase::varMetadata(jackhmmer_tbl_using_seq)

path_to_example_aln <- system.file(
  "extdata/alignment_example.afa",
  package = "HMMERutils")
alns <- Biostrings::readAAMultipleAlignment(path_to_example_aln)
jackhmmer_tbl_using_aln <- search_jackhmmer(
  alns = alns,
  dbs = "pdb",
  verbose = FALSE)
jackhmmer_tbl_using_aln
Biobase::varMetadata(jackhmmer_tbl_using_aln)
}
