% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_phmmer.R
\name{search_phmmer}
\alias{search_phmmer}
\title{Perform a phmmer search of a protein sequence against a protein sequence database.}
\usage{
search_phmmer(
  seqs,
  seq_names = NULL,
  dbs = "swissprot",
  N.TRIES = 1,
  verbose = TRUE
)
}
\arguments{
\item{seqs}{A character vector containing the sequences of the query or
any other object that can be converted to that.}

\item{seq_names}{A character vector containing the names of the sequences
of the query. By default \code{seq_names = NULL}.}

\item{dbs}{A character vector containing the target databases. Frequently
used databases are \code{swissprot}, \code{uniprotrefprot}, \code{uniprotkb}, \code{ensembl},
\code{pdb} and \code{alphafold}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{N.TRIES}{An integer specifying the number of attempts before
an error occurs.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}
}
\value{
An \code{AnnotatedDataFrame}, consisting of 2 parts, a nested DataFrame
with the search hashes, the download links of all available files and
of the HMMER page where the results are hosted, and the metadata
associated to this DataFrame. Although all available results are available
here, we recommend using the \code{extract_from_HMMER_data_tbl} function
to preprocess the data.
}
\description{
Perform a phmmer search of a protein sequence against a protein sequence database.
}
\examples{
phmmer_tbl <- search_phmmer(
  seqs = "MTEITAAMVKELRTGAGMMDCKN",
  seq_names = "1efu_B",
  dbs = "pdb",
  verbose = FALSE)
phmmer_tbl
Biobase::varMetadata(phmmer_tbl)
}
