% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sequences_to_hmmer_tbl.R
\name{add_AAStringSet_to_tbl}
\alias{add_AAStringSet_to_tbl}
\title{Join a fasta file into a tibble using "hits.name" as a common column.}
\usage{
add_AAStringSet_to_tbl(fasta, data, extension)
}
\arguments{
\item{fasta}{a AAStringSet.}

\item{data}{a tibble with "hits.name" column.}

\item{extension}{a string with the desired column name}
}
\value{
A tibble resulting from joining a fasta file using "hits.name" as a common column.
}
\description{
Join a fasta file into a tibble using "hits.name" as a common column.
}
