% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_physicochemical_properties_to_HMMER_tbl.R
\name{add_physicochemical_properties_to_HMMER_tbl}
\alias{add_physicochemical_properties_to_HMMER_tbl}
\title{Add EMBOSS-inspired theoretical physicochemical properties
using the \code{Peptides} library}
\usage{
add_physicochemical_properties_to_HMMER_tbl(HMMER_tidy_tbl)
}
\arguments{
\item{HMMER_tidy_tbl}{A \code{HMMER_tidy_tbl}.}
}
\value{
A \code{HMMER_tidy_tbl}.
}
\description{
Add EMBOSS-inspired theoretical physicochemical properties
using the \code{Peptides} library
}
\examples{
\donttest{
xml_path <- system.file(
    "/extdata/ABL_TYROSINE_KINASE.xml",
    package = "HMMERutils"
)
fasta_path <- system.file(
    "/extdata/ABL_TYROSINE_KINASE.fa",
    package = "HMMERutils"
)
data <- read_hmmer_from_xml(xml_path, fasta_path) \%>\%
    extract_from_HMMER_data_tbl() \%>\%
    add_physicochemical_properties_to_HMMER_tbl()
}
}
