% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxa_to_HMMER_tbl.R
\name{add_taxa_to_HMMER_tbl}
\alias{add_taxa_to_HMMER_tbl}
\title{Add taxonomic information to an HMMER_tidy_tbl using \code{annotate_with_NCBI_taxid}.}
\usage{
add_taxa_to_HMMER_tbl(HMMER_tidy_tbl, mode = "remote", rank_vc = NULL)
}
\arguments{
\item{HMMER_tidy_tbl}{A \code{HMMER_tidy_tbl}.}

\item{mode}{A character vector, if "local" will use a local database instead of remote resources.  You will not have to download the database but it is slower.}

\item{rank_vc}{A character vector containing the desired taxonomic ranks. If empty, all available taxonomic ranges will be retrieved.}
}
\value{
A \code{HMMER_tidy_tbl}.
}
\description{
Add taxonomic information to an HMMER_tidy_tbl using \code{annotate_with_NCBI_taxid}.
}
\examples{
xml_path <- system.file(
    "/extdata/ABL_TYROSINE_KINASE.xml",
    package = "HMMERutils"
)
data <- read_hmmer_from_xml(xml_path) \%>\%
    extract_from_HMMER_data_tbl() \%>\%
    add_taxa_to_HMMER_tbl()
}
