% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxa_to_hmmer_tbl.R
\name{add_taxa_to_hmmer_tbl}
\alias{add_taxa_to_hmmer_tbl}
\title{Add taxonomic information to a Data Frame obtained
from HMMER with a "hits.taxid" column.}
\usage{
add_taxa_to_hmmer_tbl(data, mode = "remote", rank_vc = NULL)
}
\arguments{
\item{data}{An Data Frame obtained from search function.}

\item{mode}{Either "local" or "remote". If "local" you will use a local
database instead of remote resources.  You will not have to download the
database but it is slower.}

\item{rank_vc}{A character vector containing the desired taxonomic ranks.
If empty, all available taxonomic ranges will be retrieved.}
}
\value{
A Data Frame with new taxonomic parameters.
}
\description{
Add taxonomic information to a Data Frame obtained
from HMMER with a "hits.taxid" column.
}
\examples{
githubURL <- "https://raw.githubusercontent.com/currocam/HMMERutils/master/inst/extdata/data_short.rds"
download.file(githubURL,"short_data.rds",method="curl")
data <- readRDS("short_data.rds")
add_taxa_to_hmmer_tbl(
    data = data,
    mode = "remote",
    rank_vc = NULL
)
}
