% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_hmmer.R
\name{extract_from_hmmer}
\alias{extract_from_hmmer}
\title{Split a column composed of nested lists into multiple columns.}
\usage{
extract_from_hmmer(data, column = "hits.domains")
}
\arguments{
\item{data}{Dataframe whose column is going to be splitted.}

\item{column}{Column to split.}
}
\value{
A DataFrame with columns \code{column} splitted into
several columns.
}
\description{
Split a column composed of nested lists into multiple columns.
}
\examples{
data(phmmer_2abl)
extract_from_hmmer(
    data = phmmer_2abl,
    column = "hits.domains"
)
}
