% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABL1_homologous.R
\docType{data}
\name{pairwise_identities_ABL1_homologous}
\alias{pairwise_identities_ABL1_homologous}
\title{Pairwise percentage identities for ABL1_homologous.}
\format{
A data frame with 43365 rows and 3 variables:
}
\usage{
data(pairwise_identities_ABL1_homologous)
}
\description{
A DataFrame containing 43365 pairwise_identities for
ABL1_homologous Mammalian sequences.
}
\keyword{datasets}
