% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_alignment.R
\name{pairwise_sequence_identity_heatmap}
\alias{pairwise_sequence_identity_heatmap}
\title{Plot a heatmap with the pairwise identity percentages.}
\usage{
pairwise_sequence_identity_heatmap(data, annotation = NULL)
}
\arguments{
\item{data}{A Data Frame obtained from pairwise_alignment_sequence_identity.}

\item{annotation}{A character vector with as many elements as sequences.}
}
\value{
A ggplot object
}
\description{
Plot a heatmap with the pairwise identity percentages.
}
\examples{
data(phmmer_2abl)
pairwise_alignment_sequence_identity(
    seqs = phmmer_2abl$hits.fullfasta[6:10],
    aln_type = "overlap",
    pid_type = "PID2"
) \%>\%
    pairwise_sequence_identity_histogram()
}
