% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_alignment_sequence_identity.R
\name{pairwise_sequence_identity_histogram}
\alias{pairwise_sequence_identity_histogram}
\title{Plot a histogram with the pairwise identity percentages.}
\usage{
pairwise_sequence_identity_histogram(object)
}
\arguments{
\item{object}{A pairwise_sequence_identity.}
}
\value{
A ggplot object
}
\description{
Plot a histogram with the pairwise identity percentages.
}
\examples{
data(pairwise_identities_ABL1_homologous)
pairwise_sequence_identity_histogram(pairwise_identities_ABL1_homologous)
}
