% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_alignment.R
\name{pairwise_sequence_identity_histogram}
\alias{pairwise_sequence_identity_histogram}
\title{Plot a histogram with the pairwise identity percentages (without the diagonal).}
\usage{
pairwise_sequence_identity_histogram(data)
}
\arguments{
\item{data}{A Data Frame obtained from pairwise_alignment_sequence_identity.}
}
\value{
A ggplot object
}
\description{
Plot a histogram with the pairwise identity percentages (without the diagonal).
}
\examples{
data(phmmer_2abl)
pairwise_alignment_sequence_identity(
    seqs = phmmer_2abl$hits.fullfasta[6:10],
    aln_type = "overlap",
    pid_type = "PID2"
) \%>\%
    pairwise_sequence_identity_histogram()
}
