% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmmer_from_xml.R
\name{read_hmmer_from_xml}
\alias{read_hmmer_from_xml}
\title{Read the results of a search with HMMER from an XML file and, optionally,
from a fullseq-fasta file and/or an alignment.}
\usage{
read_hmmer_from_xml(
  xml_file_paths,
  fullseq_fasta_paths = NULL,
  alignment_fasta_paths = NULL
)
}
\arguments{
\item{xml_file_paths}{A a character vector the containing filepaths to
XML file.}

\item{fullseq_fasta_paths}{If desired, a character vector containing
the paths to the fullseq-fasta file. By default, \code{NULL}.}

\item{alignment_fasta_paths}{If desired, a character vector containing
the paths to the fasta alignment file. By default, \code{NULL}.}
}
\value{
A nested DataFrame with columns \code{hits}, \code{stats}, \code{domains} and,
if selected, \code{fullseq.fasta} and \code{alignment}.

A nested DataFrame with columns \code{seqs}, \code{dbs}, \code{url}
(HMMER temporary url), \code{hits}, \code{stats}, \code{domains} and,
if selected, \code{fullseq.fasta.}
}
\description{
Read the results of a search with HMMER from an XML file and, optionally,
from a fullseq-fasta file and/or an alignment.
}
\section{Stats data structure}{

\itemize{
\item \code{nhits}: 	The number of hits found above reporting thresholds.
\item \code{Z}: 	The number of sequences or models in the target database.
\item \code{domZ}: 	The number of hits in the target database.
\item \code{nmodels} 	The number of models in this search.
\item \code{nincluded} 	The number of sequences or models scoring above the
significance threshold.
\item \code{nreported} 	The number of sequences or models scoring above the
reporting threshold
}
}

