% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hmmscan.R
\name{search_hmmscan}
\alias{search_hmmscan}
\title{Perform a hmmscan search of a protein sequence against a profile-HMM database.}
\usage{
search_hmmscan(seqs, dbs = "pfam", verbose = TRUE, timeout = 90)
}
\arguments{
\item{seqs}{A vector of characters containing the sequences of the query.}

\item{dbs}{A character vector containing the target databases. Frequently
used databases are \code{pfam}, \code{tigrfam} \code{gene3d}, \code{superfamily}, \code{pirsf} and
\code{treefam}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}

\item{timeout}{An integer specifying the number of seconds to wait for the
reply before a time out occurs.}
}
\value{
A nested DataFrame with columns \code{seqs}, \code{dbs}, \code{url}
(HMMER temporary url), \code{hits}, \code{stats}, \code{domains} and, if selected,
\code{fullseq.fasta.}
}
\description{
Perform a hmmscan search of a protein sequence against a profile-HMM database.
}
\section{Stats data structure}{

\itemize{
\item \code{nhits}: 	The number of hits found above reporting thresholds.
\item \code{Z}: 	The number of sequences or models in the target database.
\item \code{domZ}: 	The number of hits in the target database.
\item \code{nmodels} 	The number of models in this search.
\item \code{nincluded} 	The number of sequences or models scoring above the significance threshold.
\item \code{nreported} 	The number of sequences or models scoring above the reporting threshold
}
}

\section{Sequence data structure}{

The hits array contains one or more sequences. Tthe redundant sequence information will also be included.
\itemize{
\item \code{name}: 	Name of the target (sequence for phmmer/hmmsearch, HMM for hmmscan).
\item \code{acc}: 	Accession of the target.
\item \code{acc2}: 	Secondary accession of the target.
\item \code{id}: 	Identifier of the target.
\item \code{desc}: 	Description of the target.
\item \code{score}: 	Bit score of the sequence (all domains, without correction)
\item \code{pvalue}: 	P-value of the score.
\item \code{evalue}: 	E-value of the score.
\item \code{nregions}: 	Number of regions evaluated.
\item \code{nenvelopes}: 	Number of envelopes handed over for domain definition, null2, alignment, and scoring.
\item \code{ndom}: 	Total number of domains identified in this sequence.
\item \code{nreported}: 	Number of domains satisfying reporting thresholding.
\item \code{nincluded}: 	Number of domains satisfying inclusion thresholding.
\item \code{taxid}: 	The NCBI taxonomy identifier of the target (if applicable).
\item \code{species}: 	The species name of the target (if applicable).
\item \code{kg}: 	The kingdom of life that the target belongs to - based on placing in the NCBI taxonomy tree (if applicable).
}
}

\section{Domain data structure}{

The domain contains the details of the match, in particular the alignment between the query and the target.
\itemize{
\item \code{ienv}: 	Envelope start position.
\item \code{jenv}: 	Envelope end position.
\item \code{iali}: 	Alignment start position.
\item \code{jali}: 	Alignment end position.
\item \code{bias}: 	null2 score contribution.
\item \code{oasc}: 	Optimal alignment accuracy score.
\item \code{bitscore}: 	Overall score in bits, null corrected, if this were the only domain in seq.
\item \code{cevalue}: 	Conditional E-value based on the domain correction.
\item \code{ievalue}: 	Independent E-value based on the domain correction.
\item \code{is_reported}: 	1 if domain meets reporting thresholds.
\item \code{is_included}: 	1 if domain meets inclusion thresholds.
\item \code{alimodel}: 	Aligned query consensus sequence phmmer and hmmsearch, target hmm for hmmscan.
\item \code{alimline}: 	Match line indicating identities, conservation +’s, gaps.
\item \code{aliaseq}: 	Aligned target sequence for phmmer and hmmsearch, query for hmmscan.
\item \code{alippline}: 	Posterior probability annotation.
\item \code{alihmmname}: 	Name of HMM (query sequence for phmmer, alignment for hmmsearch and target hmm for hmmscan).
\item \code{alihmmacc}: 	Accession of HMM.
\item \code{alihmmdesc}: 	Description of HMM.
\item \code{alihmmfrom}: 	Start position on HMM.
\item \code{alihmmto}: 	End position on HMM.
\item \code{aliM}: 	Length of model.
\item \code{alisqname}: 	Name of target sequence (phmmer, hmmscan) or query sequence(hmmscan).
\item \code{alisqacc}: 	Accession of sequence.
\item \code{alisqdesc}: 	Description of sequence.
\item \code{alisqfrom}: 	Start position on sequence.
\item \code{alisqto}: 	End position on sequence.
\item \code{aliL}: 	Length of sequence.
}
}

\examples{
hmmscan_tbl <- search_hmmscan(
    seqs = "MTEITAAMVKELRESTGAGMMDCKN",
    dbs = "pfam",
    verbose = TRUE,
    timeout = 90
)
}
