% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hmmscan.R
\name{search_hmmscan}
\alias{search_hmmscan}
\title{hmmscan search of a protein sequence against a profile-HMM database.}
\usage{
search_hmmscan(
  seqs,
  seq_names = NULL,
  dbs = "pfam",
  verbose = TRUE,
  N.TRIES = 1
)
}
\arguments{
\item{seqs}{A vector of characters containing the sequences of the query.}

\item{dbs}{A character vector containing the target databases. Frequently
used databases are \code{pfam}, \code{tigrfam} \code{gene3d}, \code{superfamily}, \code{pirsf} and
\code{treefam}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}

\item{N.TRIES}{An integer specifying the number of trials before a time out occurs.}
}
\value{
A nested DataFrame with columns \code{seqs}, \code{dbs}, \code{url}
(HMMER temporary url), \code{hits}, \code{stats}, \code{domains} and,
if selected, \code{fullseq.fasta.}
}
\description{
hmmscan search of a protein sequence against a profile-HMM database.
}
\examples{
try(
    hmmscan_tbl <- search_hmmscan(
        seqs = "MTEITAAMVKELRESTGAGMMDCKN",
        seq_names = "seq1",
        dbs = "pfam",
        verbose = FALSE,
   )
)
}
\seealso{
\code{\link{ABL1_homologous}} for a detailed description of
the meaning of the different variables.
}
