% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hmmscan.R
\name{search_hmmscan}
\alias{search_hmmscan}
\title{Perform a hmmscan search of a protein sequence
against a profile-HMM database.}
\usage{
search_hmmscan(seq, hmmdb = "pfam", timeout = 180, verbose = FALSE)
}
\arguments{
\item{seq}{A character vector containing the sequences of the query or
any other object that can be converted to that. It can also be a URL or
the path to a FASTA file.}

\item{hmmdb}{A character vector containing the target databases. Frequently
used databases are \code{pfam}, \code{tigrfam} \code{gene3d}, \code{superfamily}, \code{pirsf} and
\code{treefam}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{timeout}{Set maximum request time in seconds.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}
}
\value{
An Data Frame containing the results from HMMER.
}
\description{
Perform a hmmscan search of a protein sequence
against a profile-HMM database.
}
\examples{
search_hmmscan(
    seq = "SWYHGPVSRNAAEYLLSSGINGSFLVRESESSPGQRSISLRYEGRVYHYRINTASDGKLYVSS",
    hmmdb = "pfam",
    verbose = FALSE
)
}
