% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hmmsearch.R
\name{search_hmmsearch}
\alias{search_hmmsearch}
\title{Perform a hmmsearch search of a protein alignment against a protein sequence database.}
\usage{
search_hmmsearch(
  alns,
  dbs = "swissprot",
  fullseqfasta = TRUE,
  verbose = TRUE,
  timeout = 90,
  alignment = FALSE
)
}
\arguments{
\item{alns}{A list of Biostrings::AAMultipleAlignment.}

\item{dbs}{A character vector containing the target databases. Frequently
used databases are \code{swissprot}, \code{uniprotrefprot}, \code{uniprotkb}, \code{ensembl},
\code{pdb} and \code{alphafold}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{fullseqfasta}{A logical, if TRUE full sequence fasta will be downloaded
as a AAStringSet. Package \code{Biostrings} must be installed in that case.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}

\item{timeout}{An integer specifying the number of seconds to wait for the
reply before a time out occurs.}

\item{alignment}{A logical, if TRUE sequence alignment will be downloaded
as a \code{AAMultipleAlignment}. Package \code{Biostrings} must be installed in that case.#'}
}
\value{
A nested DataFrame with columns \code{alns}, \code{dbs}, \code{url}
(HMMER temporary url), \code{hits}, \code{stats}, \code{domains} and, if selected,
\code{fullseq.fasta.} and \code{alignment}.
}
\description{
Perform a hmmsearch search of a protein alignment against a protein sequence database.
}
\examples{
aln <- c(
    "MTEITAAMVKELR--TGAGMMDCKN",
    "------AMVKELRESTGAGMMDCKN"
) \%>\%
    Biostrings::AAMultipleAlignment()
try(
    hmmsearch_tbl <- search_hmmsearch(
        alns = aln,
        dbs = "pdb",
        verbose = FALSE,
        timeout = 15
    )
)
}
