% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hmmsearch.R
\name{search_hmmsearch}
\alias{search_hmmsearch}
\title{Perform a hmmsearch search of a protein sequence
against a protein sequence database.}
\usage{
search_hmmsearch(aln, seqdb = "swissprot", timeout = 180, verbose = FALSE)
}
\arguments{
\item{aln}{A Biostrings::AAMultipleAlignment or a list of
Biostrings::AAMultipleAlignment.}

\item{seqdb}{A character vector containing the target databases. Frequently
used databases are \code{swissprot}, \code{uniprotrefprot}, \code{uniprotkb}, \code{ensembl},
\code{pdb} and \code{alphafold}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{timeout}{Set maximum request time in seconds.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}
}
\value{
An Data Frame containing the results from HMMER.
}
\description{
Perform a hmmsearch search of a protein sequence
against a protein sequence database.
}
\examples{
githubURL <- "https://raw.githubusercontent.com/currocam/HMMERutils/4-extract_from_hmmer/inst/extdata/alignment.fasta"
download.file(githubURL,"alignment.fasta",method="curl")
aln <- Biostrings::readAAMultipleAlignment("alignment.fasta")
search_hmmsearch(
    aln = aln,
    seqdb = 'swissprot',
    timeout = 180,
    verbose = FALSE
)
}
