% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_phmmer.R
\name{search_phmmer}
\alias{search_phmmer}
\title{Perform a phmmer search of a protein sequence against a protein sequence database.}
\usage{
search_phmmer(
  seqs,
  dbs = "swissprot",
  fullseqfasta = TRUE,
  verbose = TRUE,
  timeout = 90,
  alignment = FALSE
)
}
\arguments{
\item{seqs}{A vector of characters containing the sequences of the query.}

\item{dbs}{A character vector containing the target databases. Frequently
used databases are \code{swissprot}, \code{uniprotrefprot}, \code{uniprotkb}, \code{ensembl},
\code{pdb} and \code{alphafold}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{fullseqfasta}{A logical, if TRUE full sequence fasta will be downloaded
as a AAStringSet. Package \code{Biostrings} must be installed in that case.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}

\item{timeout}{An integer specifying the number of seconds to wait for the
reply before a time out occurs.}

\item{alignment}{A logical, if TRUE sequence alignment will be downloaded
as a \code{AAMultipleAlignment}. Package \code{Biostrings} must be installed in that case.}
}
\value{
A nested DataFrame with columns \code{seqs}, \code{dbs}, \code{url}
(HMMER temporary url), \code{hits}, \code{stats}, \code{domains} and, if selected,
\code{fullseq.fasta.} and \code{alignment}.
}
\description{
Perform a phmmer search of a protein sequence against a protein sequence database.
}
\examples{
try(
    phmmer_tbl <- search_phmmer(
        seqs = "MTEITAAMVKELRTGAGMMDCKN",
        dbs = "pdb",
        verbose = FALSE,
        timeout = 15
    )
)
}
\seealso{
\code{\link{ABL1_homologous}} for a detailed description of the meaning
of the different variables.
}
