% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_phmmer.R
\name{search_phmmer}
\alias{search_phmmer}
\title{Perform a phmmer search of a protein sequence
against a protein sequence database.}
\usage{
search_phmmer(seq, seqdb = "swissprot", timeout = 180, verbose = FALSE)
}
\arguments{
\item{seq}{A character vector containing the sequences of the query or
any other object that can be converted to that.}

\item{seqdb}{A character vector containing the target databases. Frequently
used databases are \code{swissprot}, \code{uniprotrefprot}, \code{uniprotkb}, \code{ensembl},
\code{pdb} and \code{alphafold}, but a complete and updated list is available at
\url{https://www.ebi.ac.uk/Tools/hmmer/}.}

\item{timeout}{Set maximum request time in seconds.}

\item{verbose}{A logical, if TRUE details of the download process is printed.}
}
\value{
An Data Frame containing the results from HMMER.
}
\description{
Perform a phmmer search of a protein sequence
against a protein sequence database.
}
\examples{
search_phmmer(
    seq = "MTEITAAMVKELRTGAGMMDCKN",
    seqdb = "pdb",
    verbose = FALSE
)
}
